requirements.txt
python-telegram-bot
requests
ruamel.yaml
